refactor: 접근제어자 수정, 안쓰는 import 문 정리
