+++ src/main/java/nextstep/subway/auth/application/AuthService.java
@@ -8,7 +8,6 @@ import nextstep.subway.constant.ErrorCode;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class AuthService {
