+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -19,8 +19,7 @@ public class Line extends BaseEntity {
     @Embedded
     private Sections sections = new Sections(new ArrayList<>());
 
-    public Line() {
-    }
+    protected Line() {}
 
     public Line(String name, String color) {
         this.name = name;
