+++ src/main/java/nextstep/subway/member/domain/Member.java
@@ -18,8 +18,7 @@ public class Member extends BaseEntity {
     private String password;
     private Integer age;
 
-    public Member() {
-    }
+    protected Member() {}
 
     public Member(String email, String password, Integer age) {
         this.email = email;
