+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -2,7 +2,6 @@ package nextstep.subway.member.ui;
 
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
-import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
