+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -13,8 +13,7 @@ public class Station extends BaseEntity {
     @Column(unique = true)
     private String name;
 
-    public Station() {
-    }
+    protected Station() {}
 
     public Station(String name) {
         this.name = name;
