+++ src/test/java/nextstep/subway/member/MemberRestAssured.java
@@ -89,7 +89,6 @@ public class MemberRestAssured {
         return RestAssured
                 .given().log().all()
                 .auth().oauth2(token)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().delete("/members/me")
                 .then().log().all()
                 .extract();
