+++ src/test/java/study/unit/SpringExtensionTest.java
@@ -28,7 +28,7 @@ public class SpringExtensionTest {
     @Test
     void findAllLines() {
         // given
-        when(lineRepository.findAll()).thenReturn(Lists.newArrayList(new Line()));
+        when(lineRepository.findAll()).thenReturn(Lists.newArrayList(new Line("신분당선", "bg-red-600")));
         LineService lineService = new LineService(lineRepository, stationRepository);
 
         // when
